Factor out render_gl methods since it got too long
